Extract CoreData managed model classes into seperate files
