+++ EssentialFeed/EssentialFeed/Feed Cache/Infrustructure/CoreDataFeedStore.swift
@@ -63,48 +63,3 @@ public final class CoreDataFeedStore: FeedStore {
         context.perform { action(context) }
     }
 }
-
-@objc(ManagedCache)
-internal class ManagedCache: NSManagedObject {
-    @NSManaged internal var timestamp: Date
-    @NSManaged internal var feed: NSOrderedSet
-    
-    var localFeed: [LocalFeedImage] {
-        return feed.compactMap { ($0 as? ManagedFeedImage)?.local }
-    }
-    
-    static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
-        let request = NSFetchRequest<ManagedCache>(entityName: ManagedCache.entity().name!)
-        request.returnsObjectsAsFaults = false
-        return try context.fetch(request).first
-    }
-    
-    static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
-        try find(in: context).map(context.delete)
-        return ManagedCache(context: context)
-    }
-}
-
-@objc(ManagedFeedImage)
-internal class ManagedFeedImage: NSManagedObject {
-    @NSManaged internal var id: UUID
-    @NSManaged internal var imageDescription: String?
-    @NSManaged internal var location: String?
-    @NSManaged internal var url: URL
-    @NSManaged internal var cache: ManagedCache
-    
-    static func images(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
-        return NSOrderedSet(array: localFeed.map { local in
-            let managed = ManagedFeedImage(context: context)
-            managed.id = local.id
-            managed.imageDescription = local.description
-            managed.location = local.location
-            managed.url = local.url
-            return managed
-        })
-    }
-    
-    var local: LocalFeedImage {
-        return LocalFeedImage(id: id, description: imageDescription, location: location, url: url)
-    }
-}
